refactor(api): split LabelData type to Label model
