+++ src/apis/getLabels.ts
@@ -1,17 +1,10 @@
 import { ZaloApiError } from "../Errors/ZaloApiError.js";
 import { apiFactory } from "../utils.js";
 
+import type { LabelData } from "../models/index.js";
+
 export type GetLabelsResponse = {
-    labelData: {
-        id: number;
-        text: string;
-        textKey: string;
-        conversations: string[];
-        color: string;
-        offset: number;
-        emoji: string;
-        createTime: number;
-    }[];
+    labelData: LabelData[];
     version: number;
     lastUpdateTime: number;
 };
