+++ src/models/Label.ts
@@ -1,0 +1,10 @@
+export type LabelData = {
+    id: number;
+    text: string;
+    textKey: string;
+    conversations: string[];
+    color: string;
+    offset: number;
+    emoji: string;
+    createTime: number;
+};
